Cleanup: Remove dircolors kludge from zsh config
When I first tried solarized I used the basic xresources version, which
messed up the appearance of ls output. I sort of fixed this using
dircolors to customise the appearance of ls output, but now I have
better solutions for this problem (true colour terminals, or failing
that base16-shell).