Cleanup: remove sxhkd configuration
This configuration came from a time when I was experimenting with
elementaryOS on my laptop. Configurability is not one of the project's
goals, so if I wanted custom hotkeys I needed to run sxhkd in the
background. I'm no longer interested in elementaryOS, even as a
lightweight system, so this can go.