Cleanup: remove powerline configuration files
Powerline is visually appealing and can make important information stand
out on the statusline, but I didn't like the way it introduced a soft
dependency to every app that is configured to use it. The configuration
present in this file is minimal, so I have no need to keep it.